+++ src/pint/toa.py
@@ -2404,7 +2404,7 @@ class TOAs:
         self.table.add_column(col)
 
 
-def merge_TOAs(TOAs_list):
+def merge_TOAs(TOAs_list, strict=False):
     """Merge a list of TOAs instances and return a new combined TOAs instance
 
     In order for a merge to work, each TOAs instance needs to have
@@ -2413,9 +2413,14 @@ def merge_TOAs(TOAs_list):
     .planets (i.e. whether planetary PosVel columns are in the tables
     or not).
 
+    If ``pulse_number``, [``ssb_obs_pos``, ``ssb_obs_vel``, ``obs_sun_pos``], [``tdb``, ``tdbld``]
+    columns are present in some but not all data try to put them in unless ``strict=True``
+
     Parameters
     ----------
     TOAs_list : list of TOAs instances
+    strict : bool, optional
+        If ``strict``, do not add in missing columns to facilitate merging
 
     Returns
     -------
@@ -2444,23 +2449,66 @@ def merge_TOAs(TOAs_list):
     planets = [tt.planets for tt in TOAs_list]
     if len(set(planets)) > 1:
         raise TypeError(f"merge_TOAs() cannot merge. Inconsistent planets: {planets}")
-    num_cols = [len(tt.table.columns) for tt in TOAs_list]
-    if len(set(num_cols)) > 1:
-        columns = TOAs_list[0].table.colnames
-        for i, tt in enumerate(TOAs_list[1:]):
-            extra_columns = [x for x in tt.table.colnames if not x in columns]
-            missing_columns = [x for x in columns if not x in tt.table.colnames]
-            if len(extra_columns) > 0:
-                log.warning(
-                    f"File {i+1} has extra column(s): {','.join(extra_columns)}"
-                )
-            if len(missing_columns) > 0:
-                log.warning(
-                    f"File {i+1} has missing column(s): {','.join(missing_columns)}"
-                )
-        raise TypeError(
-            f"merge_TOAs() cannot merge. Inconsistent numbers of table columns: {num_cols}"
-        )
+    # check for the presence of various computable columns
+    #   pulse_number
+    #   ssb_obs_pos, ssb_obs_vel, obs_sun_pos
+    #   tdb, tdbld
+    # if one file has these and the others don't, try to add them if strict=False
+    all_colnames = [tt.table.colnames for tt in TOAs_list]
+    has_pulse_number = np.array(
+        ["pulse_number" in colnames for colnames in all_colnames]
+    )
+    has_posvel = np.array(
+        [
+            ("ssb_obs_pos" in colnames)
+            and ("ssb_obs_vel" in colnames)
+            and ("obs_sun_pos" in colnames)
+            for colnames in all_colnames
+        ]
+    )
+    has_tdb = np.array(
+        [("tdb" in colnames) and ("tdbld" in colnames) for colnames in all_colnames]
+    )
+    has_posvel_ecl = np.array(
+        ["ssb_obs_vel_ecl" in colnames for colnames in all_colnames]
+    )
+    if has_pulse_number.any() and not has_pulse_number.all():
+        if strict:
+            log.warning("Not all data have 'pulse_number' columns but strict=True")
+        else:
+            # some data have pulse_numbers but not all
+            # put in NaN
+            for i, tt in enumerate(TOAs_list):
+                if not "pulse_number" in tt.table.colnames:
+                    log.warning(
+                        f"'pulse_number' not present in data set {i}: inserting NaNs"
+                    )
+                    tt.table.add_column(np.ones(len(tt)) * np.nan, name="pulse_number")
+    if has_posvel.any() and not has_posvel.all():
+        if strict:
+            log.warning("Not all data have position/velocity columns but strict=True")
+        else:
+            # some data have positions/velocities but not all
+            # compute as needed
+            for i, tt in enumerate(TOAs_list):
+                if not (
+                    ("ssb_obs_pos" in tt.table.colnames)
+                    and ("ssb_obs_vel" in tt.table.colnames)
+                    and ("obs_sun_pos" in tt.table.colnames)
+                ):
+                    tt.compute_posvels()
+    if has_tdb.any() and not has_tdb.all():
+        if strict:
+            log.warning("Not all data have TDB columns but strict=True")
+        else:
+            # some data have TDBs but not all
+            # compute as needed
+            for i, tt in enumerate(TOAs_list):
+                if not (
+                    ("tdb" in tt.table.colnames) and ("tdbld" in tt.table.colnames)
+                ):
+                    tt.compute_TDBs()
+
     # Use a copy of the first TOAs instance as the base for the joined object
     nt = copy.deepcopy(TOAs_list[0])
     # The following ensures that the filename list is flat
@@ -2481,23 +2529,30 @@ def merge_TOAs(TOAs_list):
         t["index"] += start_index
         start_index += tt.max_index + 1
         tables.append(t)
-    # the astropy vstack will enforce having the same columns throughout
-    # but the error messages aren't very useful.
-    # This isn't an exhaustive check
-    # (it just checks everything against the first observation)
-    # but it should be more helpful
-    columns = tables[0].colnames
-    for i, tt in enumerate(tables[1:]):
-        extra_columns = [x for x in tt.colnames if not x in columns]
-        missing_columns = [x for x in columns if not x in tt.colnames]
-        if len(extra_columns) > 0:
-            log.warning(f"File {i+1} has extra column(s): {','.join(extra_columns)}")
-        if len(missing_columns) > 0:
-            log.warning(
-                f"File {i+1} has missing column(s): {','.join(missing_columns)}"
-            )
-    nt.table = table.vstack(tables, join_type="exact", metadata_conflicts="silent")
-    # Fix the table meta data about filenames
+    try:
+        nt.table = table.vstack(tables, join_type="exact", metadata_conflicts="silent")
+    except table.np_utils.TableMergeError as e:
+        # the astropy vstack will enforce having the same columns throughout
+        # but the error messages aren't very useful.
+        # This isn't an exhaustive check
+        # (it just checks everything against the first observation)
+        # but it should be more helpful
+        message = []
+        for i, colnames in enumerate(all_colnames[1:]):
+            extra_columns = [x for x in colnames if not x in all_colnames[0]]
+            missing_columns = [x for x in all_colnames[0] if not x in colnames]
+            if len(extra_columns) > 0:
+                message.append(
+                    f"File {i+1} has extra column(s): {','.join(extra_columns)}"
+                )
+
+            if len(missing_columns) > 0:
+                message.append(
+                    f"File {i+1} has missing column(s): {','.join(missing_columns)}"
+                )
+        raise table.np_utils.TableMergeError(", ".join(message)) from e
+
+        # Fix the table meta data about filenames
     nt.table.meta["filename"] = nt.filename
     nt.max_index = start_index - 1
     nt.hashes = {}